+++ packages/sources/intrinio/src/adapter.ts
@@ -1,14 +1,25 @@
 import { Requester, Validator } from '@chainlink/ea-bootstrap'
-import { AdapterRequest, AdapterResponse, Config, MakeWSHandler } from '@chainlink/types'
+import {
+  AdapterRequest,
+  AdapterResponse,
+  Config,
+  InputParameters,
+  MakeWSHandler,
+} from '@chainlink/types'
 import { IntrinioRealtime } from './util'
 import { makeConfig, NAME } from './config'
 
-export const customParams = {
-  base: ['base', 'from', 'asset'],
+export const inputParameters: InputParameters = {
+  base: {
+    aliases: ['from', 'asset'],
+    required: true,
+    description: 'The symbol of the asset to query',
+    type: 'string',
+  },
 }
 
 export const execute = async (input: AdapterRequest, config: Config) => {
-  const validator = new Validator(input, customParams)
+  const validator = new Validator(input, inputParameters)
   if (validator.error) throw validator.error
 
   const jobRunID = validator.validated.id
@@ -39,7 +50,7 @@ export const makeWSHandler = (config?: Config): MakeWSHandler => {
   // https://github.com/intrinio/intrinio-realtime-node-sdk
 
   const getBase = (input: AdapterRequest): string => {
-    const validator = new Validator(input, customParams, {}, false)
+    const validator = new Validator(input, inputParameters, {}, false)
     if (validator.error) {
       return ''
     }