+++ packages/sources/ipfs/src/endpoint/read.ts
@@ -1,5 +1,5 @@
 import { Requester, Validator } from '@chainlink/ea-bootstrap'
-import { Config, ExecuteWithConfig } from '@chainlink/types'
+import { Config, ExecuteWithConfig, InputParameters } from '@chainlink/types'
 import { IPFS } from 'ipfs-core-types'
 import { create } from 'ipfs-http-client'
 import { CID } from 'multiformats/cid'
@@ -10,15 +10,34 @@ export { CID }
 
 export const supportedEndpoints = ['read']
 
-const customParams = {
-  cid: false,
-  ipns: false,
-  codec: false,
-  type: false,
+export const inputParameters: InputParameters = {
+  cid: {
+    required: false,
+    description: 'The CID to read. Required if IPNS is not set',
+    exclusive: ['ipns'],
+  },
+  ipns: {
+    required: false,
+    description: 'The IPNS to read. Required if CID is not set',
+    exclusive: ['cid'],
+  },
+  codec: {
+    required: false,
+    description: 'The codec to convert the data, if necessary when type is `raw`',
+    type: 'string',
+    options: ['json', 'dag-cbor'],
+  },
+  type: {
+    required: false,
+    description: 'The type of data to read',
+    type: 'string',
+    options: ['raw', 'dag'],
+    default: 'raw',
+  },
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
-  const validator = new Validator(request, customParams)
+  const validator = new Validator(request, inputParameters)
   if (validator.error) throw validator.error
 
   const jobRunID = validator.validated.id