+++ packages/sources/taapi/src/config.ts
@@ -4,9 +4,11 @@ import { Config } from '@chainlink/types'
 export const NAME = 'TAAPI'
 
 export const DEFAULT_BASE_URL = 'https://api.taapi.io/'
+export const DEFAULT_ENDPOINT = 'indicator'
 
 export const makeConfig = (prefix?: string): Config => {
   const config = Requester.getDefaultConfig(prefix, true)
   config.api.baseURL = config.api.baseURL || DEFAULT_BASE_URL
+  config.defaultEndpoint = DEFAULT_ENDPOINT
   return config
 }