+++ packages/sources/tiingo/src/config.ts
@@ -5,7 +5,7 @@ export const NAME = 'TIINGO'
 
 export const DEFAULT_ENDPOINT = 'crypto'
 export const DEFAULT_BASE_URL = 'https://api.tiingo.com'
-export const DEFAULT_WS_API_ENDPOINT = 'wss://api.tiingo.com/iex'
+export const DEFAULT_WS_API_ENDPOINT = 'wss://api.tiingo.com'
 
 export const makeConfig = (prefix?: string): Config => {
   const config = Requester.getDefaultConfig(prefix, true)