+++ packages/sources/tradermade/src/adapter.ts
@@ -32,7 +32,7 @@ export const makeWSHandler = (config?: Config): MakeWSHandler | undefined => {
     return sub
   }
   const getPair = (input: AdapterRequest) => {
-    const validator = new Validator(input, endpoints.forex.customParams, {}, false)
+    const validator = new Validator(input, endpoints.forex.inputParameters, {}, false)
     if (validator.error) return
     const base = validator.validated.data.base.toUpperCase()
     const quote = validator.validated.data.quote.toUpperCase()