+++ packages/sources/1forge/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @chainlink/1forge-adapter
 
+## 1.4.1
+
+### Patch Changes
+
+- Updated dependencies [1b54ee913]
+  - @chainlink/ea-bootstrap@1.8.0
+  - @chainlink/ea-test-helpers@1.0.1
+
+## 1.4.0
+
+### Minor Changes
+
+- e61e849e3: Refactored endpoint names to support new forex endpoints as defaults
+
+### Patch Changes
+
+- Updated dependencies [dfc4545b3]
+  - @chainlink/ea-bootstrap@1.7.1
+  - @chainlink/ea-test-helpers@1.0.1
+
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [85360aa9]
+  - @chainlink/ea-bootstrap@1.7.0
+  - @chainlink/ea-test-helpers@1.0.1
+
+## 1.3.0
+
+### Minor Changes
+
+- bea9f065: input parameters changed
+
+### Patch Changes
+
+- 54777505: added response types
+- Updated dependencies [f272a595]
+- Updated dependencies [1d55bbde]
+  - @chainlink/ea-bootstrap@1.6.0
+  - @chainlink/ea-test-helpers@1.0.1
+
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [9b3cd511]
+  - @chainlink/ea-bootstrap@1.5.0
+  - @chainlink/ea-test-helpers@1.0.1
+
+## 1.2.0
+
+### Minor Changes
+
+- 186c1621: This PR Includes changes for the following adapters: (and their from to to abilities)Polygon: from: many to: manyOpenExchangeRates: from: one to:many1Forge: from: many to: manyTradermade: from: ma
+
 ## 1.1.10
 
 ### Patch Changes